split out navbar into a component
